backend/requirements.txt

Django==4.1.7
djangorestframework==3.14.0
numpy==1.24.2
opencv-python==4.7.0.72
Pillow==9.4.0
django-cors-headers==3.12.0
